requirements.txt
gradio_imageslider
gradio==4.36.0
torch
torchvision
matplotlib
huggingface_hub
numpy
numpy-stl
flask
flask_cors
pillow